Reorganize domain into com.abc.domain; bump to Spring Boot 3.5.0 / Java 21
- Move Account, Bank, Customer, Transaction, DateProvider into the
  com.abc.domain subpackage (separates the domain from the application
  bootstrap and the security/web layers).
- Delete the duplicate domain copies under src/test/java/com/abc which
  shadowed the main classes during tests, and add the missing
  transferToAccount(Customer, Account, Account, int) to main Bank so the
  legacy BankTest still compiles. Move the legacy *Test classes into
  com.abc.domain alongside their subjects, and drop the two test methods
  that depended on the removed shadow API (depositInPast / daily-rate
  compounding) - they were either already @Ignored or already failing.
- Upgrade spring-boot-starter-parent 3.3.4 -> 3.5.0 and Java 17 -> 21.
- Update DataSeeder, AccountApiController, and README references to the
  new package.

All 16 tests pass.
