+++ src/main/java/com/abc/domain/Customer.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.domain;
 
 import java.util.ArrayList;
 import java.util.List;
