--- src/test/java/com/abc/Bank.java
@@ -1,74 +1,0 @@
-package com.abc;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Bank {
-	private List<Customer> customers;
-
-	public Bank() {
-		customers = new ArrayList<Customer>();
-	}
-
-	public void addCustomer(Customer customer) {
-		customers.add(customer);
-	}
-
-	public List<Customer> getCustomers() {
-		return customers;
-	}
-
-	public String customerSummary() {
-		String summary = "Customer Summary";
-		for (Customer c : customers)
-			summary += "\n - " + c.getName() + " (" + format(c.getNumberOfAccounts(), "account") + ")";
-		return summary;
-	}
-
-	//Make sure correct plural of word is created based on the number passed in:
-	//If number passed in is 1 just return the word otherwise add an 's' at the end
-	private String format(int number, String word) {
-		return number + " " + (number == 1 ? word : word + "s");
-	}
-
-	public double totalInterestPaid() {
-		double total = 0;
-		for(Customer c: customers)
-			total += c.totalInterestEarned();
-		return total;
-	}
-
-	public String getFirstCustomer() {
-		try {
-			customers = null;
-			return customers.get(0).getName();
-		} catch (Exception e){
-			e.printStackTrace();
-			return "Error";
-		}
-	}
-
-	public void transferToAccount(int acct1, int acct2, int ammount){
-
-		String summary = "Customer Summary";
-		for (Customer c : customers){
-			summary += "\n - " + c.getName() + " (" + format(c.getNumberOfAccounts(), "account") + ")";
-			System.out.println("num accts" +c.getNumberOfAccounts());
-		}
-
-
-	}
-
-
-
-	public void transferToAccount(Customer cust, Account fromAccount, Account toAccount, int amount) {
-		if(cust.getNumberOfAccounts()< 2){
-
-		}
-		else{
-			fromAccount.withdraw(amount);
-			toAccount.deposit(amount);
-		}		
-	}
-
-}
