--- src/test/java/com/abc/DateProvider.java
@@ -1,18 +1,0 @@
-package com.abc;
-
-import java.util.Calendar;
-import java.util.Date;
-
-public class DateProvider {
-    private static DateProvider instance = null;
-
-    public static DateProvider getInstance() {
-        if (instance == null)
-            instance = new DateProvider();
-        return instance;
-    }
-
-    public Date now() {
-        return Calendar.getInstance().getTime();
-    }
-}
